refactor: reorganize exception handling
