+++ faith/src/main/kotlin/net/thechance/faith/api/controller/AyahBookmarkController.kt
@@ -7,9 +7,7 @@ import net.thechance.faith.api.dto.bookmark.toBookmark
 import net.thechance.faith.api.dto.bookmark.toBookmarkResponse
 import net.thechance.faith.service.AyahBookmarkService
 import org.springframework.data.domain.Page
-import org.springframework.data.domain.PageRequest
 import org.springframework.data.domain.Pageable
-import org.springframework.data.domain.Sort
 import org.springframework.data.domain.Sort.Direction
 import org.springframework.data.web.PageableDefault
 import org.springframework.http.ResponseEntity
