+++ faith/src/main/kotlin/net/thechance/faith/exception/FaithExceptionHandler.kt
@@ -1,5 +1,6 @@
-package net.thechance.faith.api.controller.exception
+package net.thechance.faith.exception
 
+import net.thechance.faith.api.dto.error.ApiErrorResponse
 import org.slf4j.LoggerFactory
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
@@ -20,10 +21,10 @@ class FaithExceptionHandler {
         )
     }
 
-    @ExceptionHandler(CannotGetPrayerTimesException::class)
-    fun onCannotGetPrayerTimesError(exception: CannotGetPrayerTimesException): ResponseEntity<ApiErrorResponse> {
+    @ExceptionHandler(FailedToGetPrayerTimesException::class)
+    fun onCannotGetPrayerTimesError(exception: FailedToGetPrayerTimesException): ResponseEntity<ApiErrorResponse> {
         return createErrorResponse(
-            message = "Cannot get prayer times",
+            message = "failed to get prayer times",
             exception = exception,
             status = HttpStatus.SERVICE_UNAVAILABLE.value(),
         )
@@ -34,7 +35,7 @@ class FaithExceptionHandler {
         exception: Exception,
         status: Int
     ): ResponseEntity<ApiErrorResponse> {
-        logger.error("{}: {}", message, exception.message)
+        logger.error(message, exception)
         val apiError = ApiErrorResponse(
             status = status, message = message
         )
